+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/support/stretching/AbstractStretchingScreen.java
@@ -184,12 +184,12 @@ public abstract class AbstractStretchingScreen<T extends AbstractBaseContainerMe
     }
 
     @Override
-    public boolean mouseScrolled(final double x, final double y, final double z, final double delta) {
+    public boolean mouseScrolled(final double x, final double y, final double scrollX, final double scrollY) {
         final boolean didScrollbar = scrollbar != null
             && !minecraft.hasShiftDown()
             && !ClientPlatformUtil.isCommandOrControlDown()
-            && scrollbar.mouseScrolled(x, y, z, delta);
-        return didScrollbar || super.mouseScrolled(x, y, z, delta);
+            && scrollbar.mouseScrolled(x, y, scrollX, scrollY);
+        return didScrollbar || super.mouseScrolled(x, y, scrollX, scrollY);
     }
 
     private int calculateVisibleRows() {