+++ CHANGELOG.md
@@ -12,6 +12,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   For keybindings on Mac, CMD is now always used instead of CTRL.
 -   Missing translation key for Refined Storage controls.
 -   Missing background sprite on bulk autocrafting request buttons.
+-   Grid scrolling between grid and inventory always going to inventory.
 
 ## [3.0.0-beta.1] - 2026-04-13